[Refactor] #6 - Change folder structure
